+++ db/migrate/20170519171224_add_filters_to_exports.rb
@@ -1,0 +1,5 @@
+class AddFiltersToExports < ActiveRecord::Migration[5.1]
+  def change
+    add_column :exports, :filters, :string
+  end
+end